Isolating TreeLoader
