+++ src/GitHub.Api/GitHub.Api.csproj
@@ -238,6 +238,7 @@
     <Compile Include="UI\FileTreeNode.cs" />
     <Compile Include="UI\GitCommitTarget.cs" />
     <Compile Include="UI\TreeBuilder.cs" />
+    <Compile Include="UI\TreeLoader.cs" />
   </ItemGroup>
   <Choose>
     <When Condition="$(Buildtype) == 'Internal'">
@@ -294,6 +295,7 @@
   <ItemGroup>
     <None Include="packages.config" />
   </ItemGroup>
+  <ItemGroup />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
        Other similar extension points exist, see Microsoft.Common.targets.
